Update README.md
In this project, we will examine the portion of the kernel that addresses the management of processes and resources. We will develop a system that will allow us to create the data structures to represent processes and resources. We will also implement the oper- ations invoked by processes to manipulate other processes or to request/release various resources. In the extended version of the manager, we will also implement operations that emulate the actions of hardware interrupts. The manager will be tested using a pre- sentation shell developed as part of the project. This will allow us to test the manager without running the actual processes and using the machine’s actual hardware interrupts. Instead, the presentation shell will play the role of both the currently running process and the hardware; it will accept commands typed in by the user, and will invoke the corresponding function of the manager.